% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APaSPU.R
\name{APaSPU}
\alias{APaSPU}
\title{The function for adaptive sum of powered score (aSPU) test}
\usage{
APaSPU(betas, vcov_mat, pow = c(1:8, Inf), B)
}
\arguments{
\item{betas}{the vector (length = p) of maximum likelihood estimates; p is the number of paramters to be tested.}

\item{vcov_mat}{the p by p variance-covariance matrix of MLE.}

\item{pow}{the power vector for aSPU tests.}

\item{B}{number of simulations}
}
\value{
simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test
}
\description{
It gives the simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test.
}
\details{
the suggested power term is c(1:8, Inf)
}
\examples{
library(lme4)
data("exdat_GLMM")
fit1<-glmer(Y~.-ID + (1|ID),family="binomial",
data=exdat_GLMM,control=glmerControl(optimizer="bobyqa"))
betas<-fixef(fit1)[-1]
Cov<-vcov(fit1)[-1,-1]
APaSPU(betas,Cov,pow=c(1:8,Inf),B=1000)
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}
\references{
Zhiyuan Xu, Wei Pan (2014) Approximate score-based testing with application to multivariate
trait association analysis. Genetic Epidemiology. 39(6): 469-479.

Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014) Testing for association with multiple
traits in generalized estimation equations, with application to neuroimaging data. Neuroimage.
96:309-25
}

