% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docu.R
\docType{data}
\name{exdat_GLMM}
\alias{exdat_GLMM}
\title{an simulated example dataset for GLMM model}
\format{a 600 by 12 data table, one row for each member.}
\source{
simulated data
}
\usage{
exdat_GLMM
}
\description{
This data set contains 200 families (indicated by ID variable), each family has 3 members.
}
\author{
Zhiyuan (Jason) Xu, 2016-10-27
}

