% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.approx.aSPU.R
\name{GEE.approx.aSPU}
\alias{GEE.approx.aSPU}
\title{The function GEE.approx.aSPU compute the p-values of aSPU test
based on approximated Score vectors from GEE}
\usage{
GEE.approx.aSPU(dat, y, par, cov, groupID, corstr = "independence",
  family = "binomial", pow = c(1:8, Inf), B)
}
\arguments{
\item{dat, }{the data table for analysis}

\item{y, }{response name}

\item{par, }{variables of interest for testing}

\item{cov, }{covariates needed to be adjusted}

\item{groupID, }{group ID name}

\item{corstr, }{working correlation structure for GEE}

\item{family}{type, common ones are "gaussian" and "binomial".}

\item{pow, }{power integer candidates}

\item{B, }{simulation number to calculate P-values}
}
\value{
P-values for SPU and aSPU test based on approximated Score vector from GEE
}
\description{
The function GEE.approx.aSPU compute the p-values of aSPU test
based on approximated Score vectors from GEE
}
\examples{

data("exdat_GLMM")
gee.aSPU <- GEE.approx.aSPU(exdat_GLMM, "Y", paste0("X.",1:10), cov = NULL,
groupID = "ID",corstr="independence",family="binomial", B = 1000 )
gee.aSPU
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}
\references{
Zhiyuan Xu, Wei Pan (2014) Approximate score-based testing with application to multivariate
trait association analysis. Genetic Epidemiology. 39(6): 469-479.

Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014) A powerful and adaptive
association test for rare variants, Genetics, 197(4), 1081-95

Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014) Testing for association with multiple
traits in generalized estimation equations, with application to neuroimaging data. Neuroimage.
96:309-25
}

