% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.approxscore.R
\name{gee.approxscore}
\alias{gee.approxscore}
\title{The function gee.approxscore to construct the approximate score vector from GEE model}
\usage{
gee.approxscore(dat, y, par, cov, groupID, corstr = "independence", family)
}
\arguments{
\item{dat, }{the data table for analysis}

\item{y, }{response name}

\item{par, }{variables of interest for testing}

\item{cov, }{covariates needed to be adjusted}

\item{groupID, }{group ID name}

\item{corstr, }{working correlation structure for GEE}

\item{family}{type, common ones are "gaussian" and "binomial".}
}
\value{
Score vector, variance-covariance matrix of Score vector and p-value by Score test
}
\description{
The function gee.approxscore to construct the approximate score vector from GEE model
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}

