% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermPvs.R
\name{PermPvs}
\alias{PermPvs}
\title{PermPvs is the function to calcualte p-values for SPU test by permutation}
\usage{
PermPvs(T0s, B)
}
\arguments{
\item{T0s, }{Null statistics matrix, the SPU statistic is matrix, rows for gamma2 and columns for gamma1, where gamma2 is for traits and gamma1 is for SNPs}

\item{B}{number of simulations}
}
\value{
simulation-based p-values of the sum of powered score (SPU) with different power
}
\description{
It gives the simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test.
}
\details{
the suggested power term is c(1:8, Inf)
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}

