% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPU.mulT.R
\name{SPU.mulT}
\alias{SPU.mulT}
\title{The function for adaptive sum of powered score (aSPU) test}
\usage{
SPU.mulT(U, V, pow = c(1:8, Inf), B)
}
\arguments{
\item{U, }{Score vector}

\item{V, }{variance-covariance matrix of U}

\item{pow, }{the power vector for aSPU tests.}

\item{B, }{number of simulations}
}
\value{
simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test
}
\description{
It gives the simulation-based p-values of the sum of powered score (SPU) with different power and aSPU test.
}
\details{
the suggested power term is c(1:8, Inf)
}
\author{
Zhiyuan (Jason) Xu, Yiwei Zhang and Wei Pan
}
\references{
Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014) A powerful and adaptive
association test for rare variants, Genetics, 197(4), 1081-95

Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014) Testing for association with multiple
traits in generalized estimation equations, with application to neuroimaging data. Neuroimage.
96:309-25
}

